Added local test for CreateRecipient request
